Update tild.yaml
Updates with correct class for Tild temperature; this enables long term statistics
https://developers.home-assistant.io/docs/core/entity/sensor/#long-term-statistics

device_class	 = type of sensor
state_class = represents a measurement in present time, not a historical aggregation such as statistics or a prediction of the future. Examples of what should be classified SensorStateClass.MEASUREMENT are: current temperature, humidity or electric power.